pyproject.toml
[project]
name = "gitlab"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aws-cdk-lib>=2.174.0",
    "boto3>=1.40.74",
    "cdk-nag>=2.37.55",
    "constructs>=10.0.0,<11.0.0",
    "requests>=2.32.5",
]
